+++ cmd/otelcol/main_others.go
@@ -18,6 +18,6 @@ package main
 
 import "go.opentelemetry.io/collector/service"
 
-func run(settings service.AppSettings) error {
+func run(settings service.CollectorSettings) error {
 	return runInteractive(settings)
 }